+++ server/src/main/java/org/elasticsearch/index/engine/Engine.java
@@ -33,10 +33,7 @@ import org.apache.lucene.search.ReferenceManager;
 import org.apache.lucene.search.similarities.Similarity;
 import org.apache.lucene.store.AlreadyClosedException;
 import org.apache.lucene.util.BytesRef;
-import org.apache.lucene.util.RamUsageEstimator;
 import org.apache.lucene.util.SetOnce;
-import org.apache.lucene.util.bkd.BKDConfig;
-import org.apache.lucene.util.bkd.BKDReader;
 import org.elasticsearch.ExceptionsHelper;
 import org.elasticsearch.action.ActionListener;
 import org.elasticsearch.action.admin.indices.flush.FlushRequest;
@@ -143,8 +140,6 @@ public abstract class Engine implements Closeable {
     protected static final String FIELD_HAS_VALUE_SOURCE = "field_has_value";
     public static final long UNKNOWN_PRIMARY_TERM = -1L;
     public static final String ROOT_DOC_FIELD_NAME = "__root_doc_for_nested";
-    private static final long BKD_READER_BASE_RAM_BYTES_USED = RamUsageEstimator.shallowSizeOfInstance(BKDReader.class);
-    private static final long BKD_CONFIG_BASE_RAM_BYTES_USED = RamUsageEstimator.shallowSizeOfInstance(BKDConfig.class);
 
     protected final ShardId shardId;
     protected final Logger logger;
@@ -310,28 +305,6 @@ public abstract class Engine implements Closeable {
             }
         }
         return new ShardFieldStats(numSegments, totalFields, usages, 0L, 0L, 0L);
-    }
-
-    protected static long getPointsBytes(FieldInfos fieldInfos) {
-        long totalPointsBytes = 0;
-        for (FieldInfo fieldInfo : fieldInfos) {
-            if (fieldInfo.getPointDimensionCount() > 0) {
-                totalPointsBytes += getBKDReaderBytes(fieldInfo);
-            }
-        }
-        return totalPointsBytes;
-    }
-
-    private static long getBKDReaderBytes(FieldInfo fieldInfo) {
-        // On construction, the BKDReader constructs two byte arrays each of size packedIndexBytesLength. We add that to
-        // the base shallow size of the BKDReader and BKDConfig to get an estimate of the total memory used by the BKDReader.
-        // The IndexInputs in the BKDReader are an abstract class so we don't estimate their size due to differing concrete implementations.
-        int packedIndexBytesLength = fieldInfo.getPointIndexDimensionCount() * fieldInfo.getPointNumBytes();
-        return BKD_READER_BASE_RAM_BYTES_USED + BKD_CONFIG_BASE_RAM_BYTES_USED + 2 * byteArrayRamBytesUsed(packedIndexBytesLength);
-    }
-
-    private static long byteArrayRamBytesUsed(int length) {
-        return RamUsageEstimator.alignObjectSize(RamUsageEstimator.NUM_BYTES_ARRAY_HEADER + (long) Byte.BYTES * length);
     }
 
     /**
